refactor(cli): normalize capturer cleanup, extract shared test helpers, adopt getCapturer indirection (#410)
- Normalize browseCommand cleanup to warn stderr + propagate Close errors,
  matching the chatCommand pattern (all 4 cleanup paths now consistent).
- Extract mockCapturerInteractor, stubHistoryManager, and
  stubUnifiedHistoryProvider to shared test_helpers_test.go.
- Add getCapturer() indirection to chatCommand, mirroring browseCommand's
  override-then-delegate pattern with buildCapturer as the production path.

Architectural review follow-up — no behavioral changes to production code.
